b2s-fve-detector/requirements.txt
flask==3.1.0
anthropic==0.78.0
python-dotenv==1.1.0
werkzeug==3.1.3
